// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const SemesterStatus = {
  CONCLUIDO: 'CONCLUIDO',
  CURSANDO: 'CURSANDO',
  PENDENTE: 'PENDENTE'
} as const

export type SemesterStatus = (typeof SemesterStatus)[keyof typeof SemesterStatus]


export const SubjectStatus = {
  APROVADO: 'APROVADO',
  REPROVADO: 'REPROVADO',
  CURSANDO: 'CURSANDO',
  PENDENTE: 'PENDENTE'
} as const

export type SubjectStatus = (typeof SubjectStatus)[keyof typeof SubjectStatus]